minor mods

+++ kv-jc/src/main/java/com/kv/jc/http/json/Position.java
@@ -4,7 +4,7 @@ public class Position {
 
 	@Override
 	public String toString() {
-		return "Position [x=" + x + ", y=" + y + "]";
+		return "Position [x=" + x.intValue() + ", y=" + y.intValue() + "]";
 	}
 
 	private Double x;
